Added help_text for the username field in forms.py and removed redundant information about the name's information.
In signals.py, a try-except block was added for exception handling.

Imported 'from django.contrib import messages' in views.py.

In ArticleUpdateView, a text field was added to the form fields. Additionally, in CommentUpdateView and CommentDeleteView, a dispatch function was implemented to protect users' messages from being edited and deleted by other users.

The code in views.py was formatted using Black.